% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kaya_equation.R
\name{Kaya_equation}
\alias{Kaya_equation}
\title{Calculation of either the tonnes of Carbon instead of CO2 Emissions yearly
or the CO2 emissions}
\usage{
Kaya_equation(pop, gdp, enInt, carbInt, output_type = "CO2")
}
\arguments{
\item{pop}{numeric indicating Population size (in millions)}

\item{gdp}{numeric indicating GDP per capita (in 1000$/person)}

\item{enInt}{numeric indicating Energy intensity (in Gigajoule/$1000GDP)}

\item{carbInt}{numeric indicating Carbon intensity (in tonnes CO2/Gigajoule)}

\item{output_type}{character either "CO2" or "C"}
}
\value{
numeric representing yearly co2 emissions of a country
}
\description{
Calculation of either the tonnes of Carbon instead of CO2 Emissions yearly
or the CO2 emissions
}
\examples{
Kaya_equation(82.4, 44, 5, 0.05, "CO2")
}
\author{
Thuy Ngoc Phuong Ta
}
